Created the cast dataset and created by genre graphs
I cleaned the credits.csv and created a cast dataframe and created a csv as Phil suggested. This dataframe has the actor/actress's name, the movies they've been in, and the character they've played as, as the columns. I need help optimizing the code though. Not sure how to use the apply() method to speed up the changing row values. For the graphs.py, I added two graphs that are related to genres. I believe you need a plot.ly account in order to display the graphs. Please test them out on your machine.